requirements.txt
streamlit>=1.20
pillow>=9.0
pytesseract>=0.3.10
reportlab>=4.0
opencv-python>=4.5
numpy>=1.21
